main security 분리
