+++ src/main/java/com/todayhouse/domain/user/dao/UserRepository.java
@@ -1,4 +1,10 @@
 package com.todayhouse.domain.user.dao;
 
-public interface UserRepository {
+import com.todayhouse.domain.user.domain.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+    Optional<User> findByEmail(String email);
 }
