+++ src/main/java/com/todayhouse/domain/user/dto/response/UserLoginResponse.java
@@ -1,4 +1,14 @@
 package com.todayhouse.domain.user.dto.response;
 
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
 public class UserLoginResponse {
+    String jwt;
 }
