+++ src/main/java/com/todayhouse/global/config/WebSecurityConfig.java
@@ -1,5 +1,7 @@
-package com.todayhouse.security.config;
+package com.todayhouse.global.config;
 
+import com.todayhouse.global.config.jwt.JwtAuthenticationFilter;
+import com.todayhouse.global.config.jwt.JwtTokenProvider;
 import lombok.RequiredArgsConstructor;
 import org.springframework.boot.autoconfigure.security.servlet.PathRequest;
 import org.springframework.context.annotation.Bean;
