--- src/main/java/com/todayhouse/security/domian/user/UserRepository.java
@@ -1,9 +1,0 @@
-package com.todayhouse.security.domian.user;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByEmail(String email);
-}
